Expand variables found in bookmarks file
A bookmarks file can have entries with environment variables. This
change expands those variables before cd-ing into the resulting
directory.

This is mainly done to support hand edited files. An example could
be an entry like the following:

$HOME/backup/|bak

Which would work on different machines regardless of the user name or
$HOME location.